requirements.txt
asgiref==3.2.7
botocore==1.17.16
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.41.0
django-appconf==1.0.4
django-ckeditor==5.9.0
django-classy-tags==1.0.0
django-compat==1.0.15
django-crispy-forms==1.9.0
django-crontab==0.7.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-newsletter==0.9.1
django-tinymce==3.0.2
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
ldif3==3.1.1
oauthlib==3.1.0
pilkit==2.0
Pillow==7.1.2
psycopg2-binary==2.8.5
python-card-me==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.25.9
whitenoise==5.1.0
